Issue #11163: Reduced Size Of InputJavadocStyleNoJavadoc
