+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/javadoc/javadocstyle/InputJavadocStyleNoJavadoc1.java
@@ -15,7 +15,7 @@ tokens = (default)ANNOTATION_DEF, ANNOTATION_FIELD_DEF, CLASS_DEF, CTOR_DEF, \
 
 package com.puppycrawl.tools.checkstyle.checks.javadoc.javadocstyle;
 
-public class InputJavadocStyleNoJavadoc // ok
+public class InputJavadocStyleNoJavadoc1 // ok
 {
     public int i1; // ok
     protected int i2; // ok
@@ -110,26 +110,4 @@ class PackageClass { // ok
         void foo3() {} // ok
         private void foo4() {} // ok
     }
-
-    private class PrivateInner { // ok
-        public int i1; // ok
-        protected int i2; // ok
-        int i3; // ok
-        private int i4; // ok
-
-        public void foo1() {} // ok
-        protected void foo2() {} // ok
-        void foo3() {} // ok
-        private void foo4() {} // ok
-    }
-
-    class IgnoredName { // ok
-        // ignore by name
-        private int logger; // ok
-        // no warning, 'serialVersionUID' fields do not require Javadoc
-        private static final long serialVersionUID = 0; // ok
-    }
-
-    /**/
-    void methodWithTwoStarComment() {} // ok
 }
